wave-draw component
